Recoded Engine as state_dict/load_state_dict and run - refactored engine tests
