+++ examples/mnist/mnist_save_resume_engine.py
@@ -119,7 +119,7 @@ def run(train_batch_size, val_batch_size,
         avg_nll = metrics['nll']
         tqdm.write(
             "Training Results - Epoch: {}  Avg accuracy: {:.2f} Avg loss: {:.2f}"
-                .format(engine.state.epoch, avg_accuracy, avg_nll)
+            .format(engine.state.epoch, avg_accuracy, avg_nll)
         )
         writer.add_scalar("training/avg_loss", avg_nll, engine.state.epoch)
         writer.add_scalar("training/avg_accuracy", avg_accuracy, engine.state.epoch)
@@ -132,7 +132,7 @@ def run(train_batch_size, val_batch_size,
         avg_nll = metrics['nll']
         tqdm.write(
             "Validation Results - Epoch: {}  Avg accuracy: {:.2f} Avg loss: {:.2f}"
-                .format(engine.state.epoch, avg_accuracy, avg_nll))
+            .format(engine.state.epoch, avg_accuracy, avg_nll))
         pbar.n = pbar.last_print_n = 0
         writer.add_scalar("valdation/avg_loss", avg_nll, engine.state.epoch)
         writer.add_scalar("valdation/avg_accuracy", avg_accuracy, engine.state.epoch)
